+++ apps/service/src/features/auth/hooks/use-auth.ts
@@ -1,5 +1,5 @@
-import { useLogin } from "@/features/auth/hooks/useLogin";
-import { useLogout } from "@/features/auth/hooks/useLogout";
+import { useLogin } from "@/features/auth/hooks/use-login";
+import { useLogout } from "@/features/auth/hooks/use-logout";
 import { useGetUserInfo } from "@/pages/my/hooks/useGetUserInfo";
 import { ROUTE_PATH } from "@app/routes";
 import { useMemo } from "react";
