+++ apps/service/src/features/auth/hooks/use-logout.ts
